fix(tests): merge same-provider credential tests into existing file to fix CI shard
The standalone test file test_model_switch_same_provider_credential.py was
not found by the CI sharded runner (exit code 4 on shard 2). Move the 4
tests into the existing test_model_switch_custom_providers.py which already
has the required imports and mock fixtures.
